+++ js/eslint.config.js
@@ -6,6 +6,9 @@ export default [
   js.configs.recommended,
   ...tseslint.configs.recommended,
   {
+    ignores: ['**/node_modules/**', '**/dist/**', '**/build/**', '**/lib/**'],
+  },
+  {
     files: ['**/*.{js,mjs,cjs,ts,tsx}'],
     languageOptions: {
       globals: {